analyze-nearby-areas/requirements.txt
fastapi
uvicorn
requests
pandas
google-cloud-bigquery
